Repect address port from resolver in dialer.
The `Resolver` interface is a good abstrction for 3party discovery system. However, currently the resolved result from `Resolver` extracts only the host part, leaving the port unchanged. Some `Resolver` implementation is not happy with it. This is the fix.

It leaves an open discussion about which port should be respect: the original port returned from remote or the resolved one. Considering no placeholder can be set as dummy or dynamic in kafka servers, we can only abndon the original port, or set another config flag for optting in such behavior. Right now, we follows KISS.